Creating few changes

+++ lib/core/db/collections/allocations.dart
@@ -6,7 +6,7 @@ part 'allocations.g.dart';
 class Allocations {
   Id id = Isar.autoIncrement;
   String? name;
-  int? total;
+  double? total;
   DateTime? createdDate;
   int? totalLastAllocation;
   DateTime? lastAllocationDate;
